Increase timeouts and log test timestamps (#792)
* Configure pytest to log timestamps

We'll configure pytest to use log timestamps, which allows us to
correlate test logs with kubernetes logs.

* Bump test timeouts again

We're waiting about 5 minutes for new nodes to become ready, however
sometimes this isn't enough. It can take around one minute just to
pull the cilium image.

For this reason, we'll double the test timeouts again.